+++ include/dxc/DXIL/DxilConstants.h
@@ -516,6 +516,10 @@ enum class OpCode : unsigned {
   LinAlgMatrixReserved1 = 31, // reserved
   LinAlgMatrixReserved2 = 32, // reserved
 
+  // Debugging
+  DebugBreak = 33,        // triggers a breakpoint if debugger is attached
+  IsDebuggerPresent = 34, // returns true if debugger is attached
+
   // Group Wave Ops
   GetGroupWaveCount = 2, // returns the number of waves in the thread group
   GetGroupWaveIndex = 1, // returns the index of the wave in the thread group
@@ -580,7 +584,7 @@ enum class OpCode : unsigned {
   HitObject_TriangleObjectPosition =
       10, // returns triangle vertices in object space as <9 x float>
 
-  NumOpCodes = 33, // exclusive last value of enumeration
+  NumOpCodes = 35, // exclusive last value of enumeration
 };
 } // namespace ExperimentalOps
 static const unsigned NumOpCodeTables = 2;
@@ -1289,6 +1293,10 @@ enum class OpCode : unsigned {
   EXP_OPCODE(ExperimentalOps, LinAlgMatrixReserved0), // reserved
   EXP_OPCODE(ExperimentalOps, LinAlgMatrixReserved1), // reserved
   EXP_OPCODE(ExperimentalOps, LinAlgMatrixReserved2), // reserved
+  EXP_OPCODE(ExperimentalOps,
+             DebugBreak), // triggers a breakpoint if debugger is attached
+  EXP_OPCODE(ExperimentalOps,
+             IsDebuggerPresent), // returns true if debugger is attached
 };
 // OPCODE-ENUM:END
 #undef EXP_OPCODE
@@ -1350,6 +1358,10 @@ enum class OpCodeClass : unsigned {
   CreateNodeInputRecordHandle,
   IndexNodeHandle,
   createNodeOutputHandle,
+
+  // Debugging
+  DebugBreak,
+  IsDebuggerPresent,
 
   // Derivatives
   CalculateLOD,
@@ -1653,7 +1665,7 @@ enum class OpCodeClass : unsigned {
   NodeOutputIsValid,
   OutputComplete,
 
-  NumOpClasses = 223, // exclusive last value of enumeration
+  NumOpClasses = 225, // exclusive last value of enumeration
 };
 // OPCODECLASS-ENUM:END