+++ utils/hct/hctdb.py
@@ -1163,6 +1163,10 @@ class db_dxil(object):
         ):
             i.category = "Linear Algebra Operations"
             i.shader_model = experimental_sm
+        
+        for i in insts("DebugBreak", "IsDebuggerPresent"):
+            i.category = "Debugging"
+            i.shader_model = experimental_sm
 
     def populate_llvm_instructions(self):
         # Add instructions that map to LLVM instructions.
@@ -6195,6 +6199,8 @@ class db_dxil(object):
         )
         add_dxil_op = op_table.add_dxil_op
 
+        retvoid_param = db_dxil_param(0, "v", "", "no return value")
+
         # Add Nop to test experimental table infrastructure.
         add_dxil_op(
             "ExperimentalNop",
@@ -6637,6 +6643,28 @@ class db_dxil(object):
         )
 
         op_table.reserve_dxil_op_range("LinAlgMatrixReserved", 3)
+
+        # Debugging intrinsics
+        add_dxil_op(
+            "DebugBreak",
+            "DebugBreak",
+            "triggers a breakpoint if debugger is attached",
+            "v",
+            "nd",
+            [
+                retvoid_param,
+            ],
+        )
+        add_dxil_op(
+            "IsDebuggerPresent",
+            "IsDebuggerPresent",
+            "returns true if debugger is attached",
+            "v",
+            "ro",
+            [
+                db_dxil_param(0, "i1", "", "true if debugger is attached"),
+            ],
+        )
 
     def finalize_dxil_operations(self):
         "Finalize DXIL operations by setting properties and verifying consistency."